requirements.txt
librosa==0.7.2
matplotlib==3.2.1
numpy==1.18.4
torchaudio==0.4.0
SoundFile==0.10.2
torch==1.4.0
PyYAML==5.3.1
tabulate==0.8.7
